Don't try to call abstract methods.
When deriving from managed classes in Python don't create the final base method wrapper for abstract methods.

If an abstract method that hasn't been implemented in Python raise an Exception. Previously the base method was called, resulting in a crash.

This behaviour was noticed when trying to implement events declared on an interface when the implicit add_X and remove_X methods weren't implemented, and tests have been added to cover that case.