Fix exception pickling (#286)
* Remove Python <2.5 exception wrapping.

* Fix crashes when trying to pickle CLR exceptions.

The "args" slot of BaseException was not filled, instead we provided the
args through our __getattr__ implementation. This fails in BaseException_reduce
which depends on "args" being not NULL.

We fix this by explicitly setting the "args" slot on all CLR exception objects
on creation. This also makes tp_repr obsolete.

* Fix Python 2 compatibility.

* Remove access to obsolete message attribute.

* Fix the tests.

* Use cPickle for Python 2.

* Fix typo.