Rename several methods and add three methods
Referring to IronPython, change the name of the methods Get, Exists, SetLocal, DelLocal to GetVariable, ContainsVariable, SetVariable, RemoveVariable.
Hidden the methods SetGlobalVariable, RemoveGlobalVariable.
Add a new method 'Compile' to compile string into ast, the ast can be seen as the ScriptSource of IronPython.
Add two new methods 'Execute' and 'Execute<T>' to execute an ast and
obtain the result, corresponding to the 'Execute' method of IronPython.