Fix PY27 dll mapping in Linux/macOS
On 5062377296e3594d461ff9459b85c5eb9d660957 this was fixed for PY3
but left unchanged for PY2.

On linux/macOS the library is aliased `python2.7` while in windows its `python 27`.
Since internally it wasn't mapped to the correct library in Linux/macOS, we had to
remap it again using the dll.config file.

Closes #120