Switch to using TrueDivide instead of Divide everywhere
Fixes #802.

TrueDivide is supported in both Python 2 and 3 while Divide exists only
in Python 2. This does mean that a simple division of numbers will
behave differently in C# vs Python for Python 2, but it will at least
behave the same way in C# independent of the Python version being used.