Support changes to CLR code when domain reloads
* Adds test cases for member changes during a domain reload

* Rework the serialization of reflected types

Serialization of System.Type, MemberInfo and MethodBase is now string
based. At deserialization, use reflection to attempt to recreate the
object, which may fail safely instead of throwing a
SerializaitonException during the deserialization of the whoie data
stream. Appropriate Exceptions will now be raised when the Maybe*'s
Value property.

ClassBase objects are now de-initialized and re-initialized in Reload
mode so that it's tp_dict picks up newly added members and removed
members no longer linger.

ModuleObject clears it's cache and remove cached members from it's
tp_dict.

Minor refactoring and modernization of MethodObject and MethodBinder

* Call PyType_Modified after modifying the type

Changing a type's attribute causes problem with it's cache. Force the
type to refresh itself when modifying it.

* Refactor the member binding logic of ClassManager

So that we can use that same logic when deserializing Maybe* types

* Include info about why deserialization failed in Maybe*

* Remove TestClassReference

The test case is covered in test_domain_relaod

Co-authored-by: Benedikt Reinartz <filmor@gmail.com>
Co-authored-by: Victor Milovanov <lost@losttech.software>