Fixes a leak in ExceptionClassObject
As seen in exceptions.c (from cpython), derived exception classes must
also clean the ExceptionBase members. If we don't, we leak the Python
and C# objects set as the __cause__/Inner (amongst
others) of thrown C# exceptions.

This does not address the serializability of the PythonException class.

Also fix various PythonException (ab)uses.