Disable SourceLink VCS reading for now
Modern versions of pip use partial clones to speed the process up, which
results in a repo format v1 instead of v0 (as a "normal" clone would
do). SourceLink does not implement v1 so far, which results in a build
failure. The given environment variable disables accessing the VCS for
SourceLink during the Python package build.

Fixes #1613.