Add note about assembly locations and code block for getting assembly load errors to python.rs
Currently, `import NamespaceName` fails if dependency assemblies could not be loaded, with a normal "ModuleNotFoundError: No module named ..." error. `assembly.GetExportedTypes()` can be used to actually receive a error for one of the assemblies which were not found.

Figuring out that this is why i got ModuleNotFound took me 3 hours, and i want to help others avoid this hassle.